osm2railjson: move signal code to its own file
