+++ editoast/osm_to_railjson/src/lib.rs
@@ -2,6 +2,7 @@ mod generate_routes;
 mod generate_signals;
 mod operational_point;
 mod osm_to_railjson;
+mod signal;
 mod switch_and_buffer_stop;
 mod utils;
